# Speed up performance of range tombstones