I've discovered what seems to be a bug in the collapsing of overlapping range tombstones. Specifically, when adding a new tombstone to a collapsing RangeDelAggregator that overlapped with an existing tombstone, every tombstone to the right of the existing tombstone would be scanned over, instead of exiting the loop after the new tombstone's endpoints were inserted.

I have more testing/benchmarking to do, but the initial results are quite promising!

```
Before
    0 tombstones  14.330 micros/op 69783 ops/sec; (51 of 100 found)
 ~500 tombstones  646.740 micros/op 1546 ops/sec; (60 of 100 found)
~5000 tombstones  100793.870 micros/op 9 ops/sec; (58 of 100 found)

After
    0 tombstones  13.680 micros/op 73099 ops/sec; (51 of 100 found)
 ~500 tombstones  568.480 micros/op 1759 ops/sec; (60 of 100 found)
~5000 tombstones  5700.330 micros/op 175 ops/sec; (58 of 100 found)
```

(That's seekrandom -num=100 on a database created with fillrandom.)