@ajkr 2018-07-09
Oops I misunderstood what you meant by overlap and initially thought the bug happened every time the ranges overlap. Now I realized it's only when the endpoints overlap. Still, good find :). I will close this and review #4014 instead.